requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==3.5.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
Flask==3.1.0
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
idna==3.10
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.1.0
narwhals==1.18.4
netifaces==0.11.0
numpy==2.1.3
oauthlib==3.2.0
openai==0.28.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pipx==1.7.1
platformdirs==4.3.6
polib==1.2.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==5.4.1
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.41.1
systemd-python==234
tenacity==9.0.0
toml==0.10.2
tomli==2.0.2
tornado==6.4.2
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.3
userpath==1.9.2
wadllib==1.3.6
watchdog==6.0.0
Werkzeug==3.1.3
xlrd==2.0.1
yarl==1.17.1
zipp==1.0.0
